Make "Actions" accept floats for "x", "y" and "duration" properties
* Use floats as type for x, y and duration so that it doesn't break when a float is provided in Actions JSON
* When floats are provided in Actions JSON, round the values and provide warning to user if rounded value is different from provided value
* Also, add 'warn' as a log type (along with 'info', 'debug' and 'error')

+++ espresso-server/app/src/androidTest/java/io/appium/espressoserver/lib/helpers/w3c/adapter/W3CActionAdapter.java
@@ -18,15 +18,15 @@ public interface W3CActionAdapter {
     void keyUp(W3CKeyEvent keyUpEvent) throws AppiumException;
 
     void pointerDown(int button, String sourceId, PointerType pointerType,
-                     Long x, Long y, Set<Integer> depressedButtons,
+                     Float x, Float y, Set<Integer> depressedButtons,
                      KeyInputState globalKeyInputState) throws AppiumException;
 
     void pointerUp(int button, String sourceId, PointerType pointerType,
-                     Long x, Long y, Set<Integer> depressedButtons,
+                   Float x, Float y, Set<Integer> depressedButtons,
                      KeyInputState globalKeyInputState) throws AppiumException;
 
     void pointerMove(String sourceId, PointerType pointerType,
-                     long currentX, long currentY, long x, long y,
+                     float currentX, float currentY, float x, float y,
                      Set<Integer> buttons, KeyInputState globalKeyInputState) throws AppiumException;
 
     void pointerCancel(String sourceId, PointerType pointerType) throws AppiumException;
@@ -45,7 +45,7 @@ public interface W3CActionAdapter {
 
     int pointerMoveIntervalDuration();
 
-    void sleep(long duration) throws AppiumException;
+    void sleep(float duration) throws AppiumException;
 
     void waitForUiThread();
